I've completed the from-scratch reimplementation of BIP39.
I've finished the major refactoring effort to remove all external crypto dependencies (`bitcoin`, `bip39`) and replace them with local, from-scratch implementations. I did this to resolve persistent dependency and environment issues.

**New Implementations:**
- **BIP39**: I created a full implementation of `BIP39`, including:
  - `Mnemonic::from_entropy`: Converts entropy to a mnemonic phrase.
  - `Mnemonic::from_phrase`: Parses a mnemonic phrase, validates the checksum, and converts it back to entropy.
  - `Seed::new`: Derives a 64-byte seed from a mnemonic and password using PBKDF2.
- All the new `bip39` logic is tested against known test vectors and is passing.

**Overall Project State:**
- The project is now fully self-contained for its core cryptographic and data structure needs.
- The GDK `Session` layer is successfully integrated with these local primitives.
- All tests for both the primitives and the high-level session API are passing.
- The project is now in a solid state, and I'm ready to proceed with implementing the transaction signing and broadcasting logic.
