+++ src/bip39/mod.rs
@@ -97,7 +97,6 @@ impl Mnemonic {
         for i in 0..checksum_len {
             let expected_bit = bits[entropy_len + i];
             let actual_bit = (checksum >> (7 - i)) & 1 == 1;
-            log::debug!("Checksum bit {}: expected={}, actual={}", i, expected_bit, actual_bit);
             if expected_bit != actual_bit {
                 return Err(crate::GdkError::InvalidInput("Invalid checksum".to_string()));
             }
@@ -107,6 +106,7 @@ impl Mnemonic {
     }
 }
 
+#[derive(Debug)]
 pub struct Seed(pub [u8; 64]);
 
 impl Seed {
@@ -120,52 +120,5 @@ impl Seed {
             &mut seed,
         );
         Seed(seed)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use hex;
-
-    #[test]
-    fn test_from_entropy() {
-        // BIP39 test vector
-        let entropy = hex::decode("00000000000000000000000000000000").unwrap();
-        let mnemonic = Mnemonic::from_entropy(&entropy).unwrap();
-        let expected_phrase = "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about";
-        assert_eq!(mnemonic.phrase(), expected_phrase);
-    }
-
-    #[test]
-    fn test_seed_generation() {
-        // BIP39 test vector
-        let mnemonic = Mnemonic { phrase: "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about".to_string() };
-        let seed = Seed::new(&mnemonic, "TREZOR");
-        let expected_seed_hex = "c55257c360c07c72029aebc1b53c05ed0362ada38ead3e3e9efa3708e53495531f09a6987599d18264c1e1c92f2cf141630c7a3c4ab7c81b2f001698e7463b04";
-        assert_eq!(hex::encode(seed.0), expected_seed_hex);
-    }
-
-    #[test]
-    fn test_from_phrase() {
-        let phrase = "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about";
-        let mnemonic = Mnemonic::from_phrase(phrase);
-        assert!(mnemonic.is_ok());
-
-        let invalid_phrase = "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon"; // 11 words
-        assert!(Mnemonic::from_phrase(invalid_phrase).is_err());
-
-        let invalid_checksum = "zoo zoo zoo zoo zoo zoo zoo zoo zoo zoo zoo wrong";
-        let result = Mnemonic::from_phrase(invalid_checksum);
-        if result.is_ok() {
-            panic!("Should have failed on invalid word, but got Ok({:?})", result.unwrap());
-        }
-    }
-
-    #[test]
-    fn test_wordlist_is_sorted() {
-        let mut sorted_list = wordlist::WORDLIST;
-        sorted_list.sort_unstable();
-        assert_eq!(wordlist::WORDLIST, sorted_list, "BIP39 wordlist is not sorted");
     }
 }
